fix: MyPy batch 3 (100 errors from 25 files)
Fixed type checking errors across framework modules:
- victor/framework/coordinators/yaml_coordinator.py: Fixed WorkflowStreamContext params and stream method
- victor/framework/di/container.py: Added cast for scoped instances, fixed Type parameters
- victor/framework/di/migration_example.py: Added return type annotations (-> None)
- victor/framework/discovery.py: Fixed EntryPoints type annotation with Union
- victor/framework/escape_hatch_registry.py: Removed redundant casts
- victor/framework/graph_cache.py: Added type parameters to StateGraph/CompiledGraph, fixed int assignment, removed unused ignore
- victor/framework/hitl/session.py: Added type parameter to Awaitable, annotated gate_counts
- victor/framework/ingestion/chunker.py: Annotated current_chunk and current_batch
- victor/framework/parallel/executor.py: Fixed TaskInput type parameters, added AsyncGenerator return, fixed union-attr with None checks
- victor/framework/patterns/miner.py: Added cast for return value, annotated role_counts
- victor/framework/patterns/protocols.py: Added return type annotation, fixed create_coordinator call
- victor/framework/resilience/: Added type ignore for attr-defined, fixed non_retryable_exceptions type
- victor/framework/rl/learners/prompt_template.py: Fixed dict.get() tuple key usage
- victor/framework/rl/rl_coordinator_enhanced.py: Added cast for return, added Experience type parameters
- victor/framework/service_provider.py: Implemented abstract methods, removed redundant casts, fixed Agent call
- victor/framework/services/lifecycle.py: Added type ignore for ServiceRegistry, fixed Task type, added return type
- victor/framework/shim.py: Added casts for return values
- victor/framework/task/complexity.py: Changed TaskTypeClassifier to Any
- victor/framework/tool_dependency_deprecation.py: Added type parameters to DeprecatedConstantDescriptor and tuple
- victor/framework/tool_packs.py: Added cast for return value
- victor/framework/validation/pipeline.py: Fixed return type to bool
- victor/framework/validation/validators.py: Fixed type annotations, removed unused ignores, fixed pattern assignment
- victor/framework/validation/yaml.py: Added return type annotation, fixed optional field handling

+++ victor/teams/ml/team_member_selector.py
@@ -300,15 +300,15 @@ class TeamMemberSelector:
         features.append(1.0 if member.is_manager else 0.0)
         features.append(int(member.max_delegation_depth))
         features.append(1.0 if member.memory_enabled else 0.0)
-        features.append(int(len(member.expertise)))
-        features.append(int(len(member.backstory)))
+        features.append(float(len(member.expertise)))
+        features.append(float(len(member.backstory)))
 
         # Expertise overlap
         if task_features.required_expertise:
             overlap = len(set(member.expertise) & set(task_features.required_expertise))
-            features.append(int(overlap / len(task_features.required_expertise)))
+            features.append(float(overlap / len(task_features.required_expertise)))
         else:
-            features.append(0)
+            features.append(0.0)
 
         # Role encoding (one-hot)
         roles = ["planner", "researcher", "executor", "reviewer", "tester"]
